[server] Introduce sidecar container which end when other containers end
Some container in a pod are utility container that should just end when
the job ends. Until now all container had to terminate for a job to be
done. This change introduces sidecar container which are terminated
automatically when all non-sidecar container are done. Also, sidecar
container automatically get their name prepended to their log output.